Added motion system

+++ ECS/Components/Velocity.cs
@@ -6,10 +6,10 @@ namespace ECS.Components
     {
         public string Name { get; set; }
         public string Direction;
-        public float Speed;
+        public int Speed;
         public bool IsMoving;
 
-        public Velocity(float speed, string direction)
+        public Velocity(int speed, string direction)
         {
             this.Name = "velocity";
             this.IsMoving = false;
